requirements.txt
setuptools==54.0.0
Cython
matplotlib==3.2.2
opencv-python>=4.1.2
pillow==8.0.1
easydict==1.9
PyYAML==5.3
scipy==1.4.1
numpy>=1.18.5
requests==2.25.1
tensorboard>=2.2
tqdm>=4.41.0
wandb
seaborn>=0.11.0
pandas
torch>=1.7.0
torchvision>=0.8.1
